+++ backend/src/common/logging.py
@@ -10,8 +10,6 @@ import structlog
 
 structlog.configure(
     processors=[
-        structlog.stdlib.filter_by_level,
-        structlog.stdlib.add_logger_name,
         structlog.stdlib.add_log_level,
         structlog.dev.ConsoleRenderer() if __debug__ else structlog.processors.JSONRenderer(),
     ],